(QENG-6081) Remove beaker dependecy
In order to move beaker over to CI.Next, this must be removed to avoid
circular dependecies. This was only required because of when the
custom hypervisor API became available. At this point, that version
of beaker is 2+ years old and we can almost safely say no one
should be using version <2.9.0.